requirements.txt
flask
openai-whisper
gTTS
textblob
